feat: cliengt/todo is dynamic to view
