+++ client/app/interfaces/TaskType.ts
@@ -1,0 +1,4 @@
+type TaskType = {
+    check: boolean,
+    task: string
+}
