+++ client/app/interfaces/TodoType.ts
@@ -1,4 +1,4 @@
-interface Todo {
+type TodoType = {
     category_id: null | number;
     flat_share_id: number;
     id: number;
